requirements.txt
docker-py
flask
redis
